requirements.txt
asgiref==3.5.2
Django==4.0.4
django-filter==22.1
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2==2.9.9
sqlparse==0.4.2
whitenoise==6.2.0
pint
